Rename _distinfo module to distinfo
